refactor(Typewine)add buton for nav
