+++ src/components/PageScan.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 function PageScan() {
   return (
-    <p>hello tonton sommelier </p>
+    <p>hello tonton sommelier</p>
   );
 }
